Update pong.py
Added collision logic. There is a bug when the ball speeds up enough which causes the ball to ride along the top of the window. Future updates may include a maximum and minimum ball.dx and ball.dx to fix this. There is also a bug where the ball sometimes phases through the center of the paddles. I am not sure why this happens, but I will try to work on it and fix it.